chore: fix lint errors in skill tree
